Updates in State and Variable classes.
Decided to create a singleton "GlobalState" as the single vector of
Variables in the system. The other State "instances", the ones which
will be contained in the Traials, can be validated agains GlobalState.
Read in-code TODO within the class.

GlobalState will also provide the encode/decode routines between the
symbolic and concrete state space representations.

Also: MuParser library can work with shorts, unsigned and longs as
base type. This avoids the need to convert between floating points
and integral values for transitions updates and state enconde/decode.
The only "forbidden type" seems to be int.