Variable tests close to its end.
There is one big fat problem still: template dependent types.

For the template class VariableSet<T_> we offer the possibility of copy/move
construction from any container with the same internal data type T_.
That means that copy/move ctors are declared as template members of this
template class. The template parameter of these ctors is a container class,
named "Set_" in VariableSet.h, which in turn is a template whose only parameter
is its interal data type, say TT_.

We need to learn how to declare that the ctors template parameter, i.e. the
container class "Set_", is dependent on T_, which is the template parameter
of the template class VariableSet<T_>. In short, we need to declare the
dependece of "Set_" from T_, to then impose the restriction TT_ == T_.

This is extremely painful for my poor Gehirn.