Major changes in Traial and TraialPool.
In Traial the changes were internal: the ordered Timeouts vector is now
implemented with an indices vector, specifying the indices in the clocks_
vector for a time-increasing view of the Clocks.
This was Raúl's idea, and makes copy/move construction trivial.

In TraialPool we now keep unique_ptrs to the available Traials,
since we couldn't return/recover references easily with move semantics.

There's a TODO in the middle of the TraialPool.h pointing at where to continue
with these modifications.