Trying to upload Transition creation example for Raúl.
We switched to this task momentarily, only to offer Raúl something to
work with. Finish as quick as possible and go back to the design issues
found for the Model family class.

Problem: seems like all builtin functions from MuParser are offered only
when MUP_BASETYPE is floating point, and we want to user integral types.

Solution: offer traditional integer functions (min, max, ...) as user
customized functions added to the MathExpression.

TODO: find generic way to offer min() and max(), that is, with an
arbitrary number of parameters. Read in-place comment in MathExpression.cpp