Extended Precondition tests -- Changes in Postcondition.
Read the Postcondition class doxygen description in the header,
where the new expected behaviour of the operator() is described.

Biggest modification is, we no longer expect a "x = expr1 , y = expr2"
list of updates, since that accumulates changes. e.g. if 'x' appears in
'expr2', then the value of 'x' used to evaluate the new value of 'y'
is currently the result of 'expr1', and we want the opposite behaviour.

We will need to take an unnamed list of comma separated updates,
"expr1, expr2" in the example above. Parallely we'll keep an array/list
of variable names, "x,y" in the example above, positionally indicating
which update corresponds to which variable in the global state.