New system model construction philosophy.
We rather allow Raúl to define the network incrementally.
For that we need a couple of callbacks, since global info
(such as the position of the clocks in a global array)
won't be known at the time of creation of transitions and
single module instances.

Some information, such as the "resetClocks_" Bitflag of a transition,
can only be known when the global info is available. For those purposes
there's a "christalize()" callback in Transition, and a "mark_added()"
callback in ModuleInstance. They are to be invoked when the ModuleInstance
is definitely added to the network with "add_module()"

Current state of affairs won't compile, not even if shouted at,
threatened, or thrown down a bridge.
Several in-place TODO comments mark the current code holes.
Search and implement all those functions.