Implementing ModuleNetwork::seal()
We need a version of the callbacks which accepts, instead of
an std::unordered_map, an std::function<size_t(std::string)>
to map the variables to their global positions.
This way the ModuleNetwork can pass the member function of its gState
to the ModuleInstances' seal(), which is already built by the time
all modules have been added to the network.

TODO_1: implement ModuleInstance::seal() taking this std::function<>
TODO_2: implement the analogous for Transition::callback()