Updates in ModuleNetwork and Traial.
ModuleNetwork::seal() now takes as argument the list of initial clock names,
i.e. those whose values have to be reset at the start of any simulation.
Traial offers now also an initialize() routine, to reset its internal state
and initial clock values when a simulation starts.

There's a description of the model to simulate in the ModelNetwork test.
It was left as a comment in the corresponding test within the source file
TADs_instantiation.cpp.