New design for higher abstraction levels.
ModelSuite will be the uppermost class (a singleton for a change)
Uploading with this commit sketches for its main processing functions.

ModelNetwork::simulation_step() still has to be implemented.
The main barrier is the so far undesigned member function
"Event inspect(const StateInstance&) const".

The main idea behind "inspect(...)" is to review a Traial.state
after a whole simulation_step(), and report the newly observed events,
such as importance threshold crossings or visits to the rare set.
It probably needs an ImportanceFunction& as argument as well.