Traial inspection will exist in SimulationEngine.
Simulation engines will hold the semantics of the current
simulation strategy, and thus know which events are relevant
for stopping the current simulation step.

The call stack should be:
 0. ModelSuite::estimate(...) calls:
 1. SimulationEngine::simulate(property, impFun, effort) calls:
 2. ModuleNetwork::simulation_step(traial, engine*) calls:
 3. SimulationEngine::eventTriggered(traial)

So in step 1., wherein the engine invokes the simlation_step()
from the model, it passes its 'this' pointer as second argument.
The model is thus freed from interpreting the end of the simulation step,
since this is checked by the engine itself after each synchronized jump,
by means of a call to engine.eventTriggered()

We're currently held back by the lack of a Property class. Several
things depend on the property, such as the importance assessment,
and also the interpretation of the effort (StoppingCondition).

TODO_0: define Property class ASAP
TODO_1: define StoppingCondition class
TODO_2: implement the class family "SimulationEngine"