Revising ImportanceFunction class family -- UNFINISHED.
Besides the ImportanceFunction name there are assessment strategies.

The "name" of an importance function will uniquely determine the way
it stores the importance information internally, and thus which derived
class it implements, but nothing else.
Possibilities are "concrete_coupled", "concrete_split", "symbolic_coupled"...

Independently, when calling "assess_importance()" the assessment "strategy"
is also requested now. This tells the function, whichever it is, how to
measure the importance for the given model and property.
Possibilities are "", "auto" and "adhoc".

See ImportanceFunction::assess_null_importance() and
ImportanceFunctionConcreteCoupled.cpp:53 for continuation points.