Struggling against mischievous simulations.
The thing is moving... but not so well.
We can launch and finish the first "warm up" simulation,
but the second attempt keeps jumping synchronously without reaching
neither a "goal" nor a "stop" state. That's odd.

Notice we've changed the model. Basically we've swapped some clock
distributions to make it more likely for the second queue to fill up.

Currently studying Traial's state in ModuleNetwork::simulation_step()