Updates in time simulation.
It works, but there's an issue with the signal handling.
Apparently we can't use lambdas with captures for the handling,
because then the function pointer doesn't match the C API.
This is a problem, since we should tell the engine that it has been
interrupted from within the handler, in order to make it stop.

There's a TODO in ModelSuite.cpp:387 with an SO workaround.