Reviewing event mark/inspect for simulations.
When using concrete importance functions, the events can be marked
along the concrete states importance. Inspecting that is just a XOR,
much faster than calling property.is_whatever(StateInstance&)

Those changes are being implemented, but there's something yet
unresolved: passing the proper "event_triggered()" member function
of the SimulationEngine, to the "simulation_step()" method of the
ModuleNetwork. Currently we're passing a reference to the whole
engine, and we want to pass only the proper "event_triggered()"
function that needs to be called after each synchronized jump.

Read: http://stackoverflow.com/a/12662961
      http://stackoverflow.com/q/7218519
      http://stackoverflow.com/a/2402607
      http://stackoverflow.com/q/14677997