Member function pass done -- Updates in auto ifun.
After some benchmarking (in ASUS laptop and in Mendieta)
the C++ NSMF API was chosen for passing the "non-static member function"
of the SimulationEngineNosplit to the ModuleNetwork::simulation_step()

A few other updates are also included in this commit, mostly
modularization inside ImportanceFunctionConcrete.cpp.

Continue with the third and last step of the automatic importance function
construction: BFS to compute the distance of every concrete state to the set
of rare states.