Major modifications in threshold building.
The building of the thresholds was removed from within the computation
of the importance function. It is now a routine offered by the ModelSuite.

Base API for ThresholdsBuilder familiy class is being changed,
the generic template function 'tune()' is to be replaced by a more
OOP approach.
Continue from ModelSuite::build_thresholds() in the source file,
where the usage of a ThresholdsBuilder is being defined.