Interruption handling with pritty print.
Even though the current code is far from reentrant we will greatly
benefit from the printed info in the dark days to come.

Currently there's no time to do things POSIX-compliant,
but when the logging mechanism is implemented we should try
to make "interrupt_print()" and the SignalSetter's lambdas
a bit more safe as signal handlers.