Fixing RESTART simulations to match NOSPLIT.
We could replicate the results for the current model,
but there's an unjustifiable numerical trick in the middle.
When weighing the rareCounts in the function
SimulationEngineRestart::transient_simulations()
we need to multiply the i-th count by "splitsPerThr ^ (numThr+i+1)"
The problem is the final "+1" which I can't explain.

If we take it out we get half of the estimation incurred by the NOSPLIT
engine, which makes some sense since splitsPerThr == 2.
Ask Pedro about this and test with more models.