Fixed "+1" in RESTART's rareCount's weighing.
With Pedro and Raúl we solved the mistery: the rare event laid
on a threshold level-up and was counted before the splitting
took place. Thus we were counting "+1" rare event in the upper
threshold level, when that was actually worth "+splitsPerThresholds"
rare events on that level.

TODO: implement symbolic or "on-the-fly" ImportanceFunction derived
      class, and also the ad-hoc importance strategy to use it with.