Ad hoc importance assessment works but there's a bug.
The bug should be in the RESTART simulation routine.
It was revealed by chance when testing "adhoc" assessment,
which by the way is doing just great.

In short: when we get more thresholds there seems to be an underestimation.
So far we'd only had three threshold levels; now we got four and the
values estimated by the RESTART engine converge to 0.08, whereas the
NOSPLIT engine converves to 0.095 (as it always has).