Implementation of ImportanceFunctionConcreteSplit -- UNFINISHED.
The idea is to have a user-defined function, just like in the ad hoc
strategy, which tells us how to combine the importance obtained from
the modules composing the system. The function could be a simple operand
which would be applied to all modules (e.g. "max", "+"), or a fully
defined function with the modules names (e.g. "5*Queue1+Queue2")

How to compute the importance in each ModuleInstance is yet undefined.