Split ifun is incomplete but estimations test compiles.
To finish the importance assessment in ImportanceFunctionConcreteSplit
we have to wait for Raúl to give us the "property simplification" method.
We discussed it already and it seems simple, I hope it's ready soon.

In the meanwhile we can play around with the tandem queue model.
Make it compile with the parser and run some tests with what we have now.