Internal changes in {pre,post}conditions expressions.
Moved "MuParser expression formatting" entirely into MathExpression class.
As before, Preconditions whose expression is "" or "true" will always
evaluate as true.
There's now also the possibility to have "true" and "false"
in Postcondition updates, which will be changed to the corresponding
MuParser expressions.

Also modified a variable in the tandem queue model to be boolean.
TODO: talk with Raúl about the Parser warnings when we compile the model.