Changes to cope with MuParser logical format.
MathExpression was modified to add the logical not operator
('!', which surprisingly isn't a MuParser builtin) and to
change '&' and '|' for '&&' and '||' respectively.
This way the logical interface with MuParser should work.
It doesn't.

Things compile but simulation tests are broken.
Fix that NOW! Then continue developing the estimation tests.

Also modularized a couple of string utils in the new "string_utils.h" header.