Major changes due to split importance function.
Raúl finished his "reduceProperty()" function opening the way
to complete the implementation of ImportanceFunctionConcreteSplit.

Problem was, the whole project's design was customized for operating
with the system's global state. Both Properties and MathExpressions
evaluations of "local module states" were not possible, since we always
stored the global state's Variables position.

Several changes were made to allow this "local evaluation" of formulas.
They operate with State<> objects and are slower than the ones previously
designed for StateInstace objects. It's fine since they are needed to
build the importance functions only. Simulations are always carried out
with the old, fast versions.

Things compile but fail to run.
  TODO_0: fix project and run simulation tests again.
  TODO_1: make sure ImportanceFunctionConcreteCoupled still works fine.
  TODO_2: test quickly ImportanceFunctionConcreteSplit.
  TODO_3: go back to work with estimation tests.