Simulation tests work again.
The SIGSEGV was due to lack of initialization of a vector.
Everything seems to work as before, now using the new MathExpression
evaluation of State<> instances for the construction of the
importance functions.

ImportanceFunctionConcreteSplit could not be used in simulation tests
because it requires an active parser instance.
We'll play with it now in estimation tests.