Estimations fixed but yield wrong results.
The bug was due to the lack of sign of the floating point zero.
Now we consider a null estimate enough a reason to enlarge the
batch size.

However, the estimation tests ran for ad hoc and automatic (coupled)
importance function yield the wrong number. The computed probability
of overflow is roughly twice as big as it should for the tandem queue
with capacity 8.

In other news, estimations with the split automatic ifun still
don't fly. An assert fails regarding the number of rare states
needed for the backwards BFS analysis.

TODO: focus on fixing the split importance function;
      on monday talk to Pedro about the wrong estimates.