Thresholds building fixed -- Estimation tests run completely.
All importance functions and simulation engines are currently
running and producing roughly the same estimate for the test model:
the tandem queue with capacity '8' and property "q2>0 U lost".

Such estimate is however wrong w.r.t. our published results with
bluemoon. The value obtained now is about twice bigger than the one
reported in the article, e.g. we get ~1e-5 and it should be ~5e-6.

Talk to Pedro.