Finished SMC thresholds builder - Problem was elsewhere.
The new ThresholdsBuilder is complete and was briefly tested.
It seems to be doing right but we should revise the code once again,
since the theory is non-trivial and there are missing docstrings left.

Anyway, the problem was actually in the atm_queue.sa model,
where the subexpression "buf < 0" was avoiding buffer enqueueing.
That was fixed and we have to test how does it move now.
We're also translating the "queue with breakdowns" of Kroese & Nicola
from bluemoon's model.

TODO_0: review ThresholdsBuilderSMC code to make sure it's ready for use
TODO_1: experiment a bit with the fixed atm_queue.sa model, gain confidence
        in the way we're doing things (it's unnecessary to get the results)
TODO_2: finish translating bluemoon's "queues_with_brakdowns.prism"
        and simulate on it -- For this one we do need the right results