Both thresholds builders looking good now.
AMS was updated in the simulations stopping condition: we now also stop
the peak simulations when they reach the maximum ImportanceValue.
This resulted in some more thresholds found for the ATM model.

SMC was also updated in the simulations stopping condition for the
choosing of the initial states distribution. In this case we added
an upper bound for the number of jumps, after which we give up and
start again from another initial state.

Both classes presented aliasing bugs which we hope are now fixed.
Some time was spent running the ATM model and the results are
satisfactory. Closing this task.

TODO: translate Nicola & Kroese's breaking queues model from bluemoon.