Bugfix in SimulationEngineRestart.
The tracking of a Traial's depth and current level was wrong, resulting
in early kills. New attribute added to Traial class to register the number
of threshold levels up-crossed, keeping that info independently from the
'depth' and 'level' of the Traial.

These changes were tested on the two working models (queues_with_breaks
and tandem_queue) successfully. The new model (3tandem_queue) seems to be
running properly now, but there are no results yet to testify the model's
validity.

TODO_0: reproduce the paper's results with our 3tandem_queue model.
TODO_1: implement the oil pipeline model /or/ the database model.