Refactored thresholds storage inside importance functions.
We needed the importance information to prevail after choosing
the thresholds, to choose new thresholds again without recomputing
all the states' importance again.

This meant ImportanceFunctionConcreteCoupled can't replace its importance info
with the thresholds levels, and that cascaded into a great pulling up of the
threshold-related member functions in the ImportanceFunction class family.
Now everything lies at the highest abstration level: the asbtract base
ImportanceFunction class.

Also a couple of bugs were fixed, e.g. we now always mantain the real
ImportanceValue extreme values stored in the importance function,
and when asked for the "max_value()" after computing the thresholds
we simply translate on the fly.