New thresholds builders.
The idea is to implement the hybrid system, where some adaptive builder
starts selecting the thresholds and when it fails, we resource to some
deterministic choice of importance values.
We need this because even SMC failed in Mendieta for the hard models.

Currently implementing the basic "fixed thresholds choosing" by means of
the class ThresholdsBuilderFixed. The proposed ThresholdsBuilderHybrid
would use, say, SMC until it fails, and then resort to this fixed variant.

Simulations fail when ThresholdsBuilderFixed chooses to set one threshold
per ImportanceValue (i.e. the thinnest granularity level)
It may have something to do with the importance of the system's initial
state, from which we should start the thresholds selection. No thresholds
should ever be set below the initial state.