Finished ThresholdsBuilderFixed + Bugfix in TraialPool.
The bug found in the TraialPool was actually patched, not fixed,
since a permanent solution may require some heavy refactoring.
The problem is deeply rooted in the use of move semantics to handle
the Traial instances. Proper documentation was left in the source code
of the class, namely inside TraialPool::ensure_resources()