Bug tracking is pointing to wrong clock values.
At some point we get stuck in a state where the clocks start taking
increasingly negative values. Find the cause and we may fix this bug.

To reproduce estimate the Rate property for the tandem queue with
capacity 9, for default splitting and thresholds technique.
At some point in the 18th batch iteration we visit the state (8,0,0)
and some clock reaches a negative value below -10.
The behaviour from then on is to reset 'clkArr' and decrement 'clkPass'
and 'clkExit' indefinitely.