Two bugfixes.
The misterious wrong clock values were caused by two clocks which
ended up on the top of the "next_timeout" stack with equal values.
When the first was attended then the second was also reset to 0.
Asking for the "next not null clock" afterwards resulted in any
(wrong) clock returned, changing the semantics of the whole model.

Pedro found a workaround: expired clocks are now given a negative
value (namely -100 was chosen) and non-expired clocks can also have
a null value. Thus we now ask for the 'next non-negative clock'.

With that change, if the previous situation happens, the first
attended clock is set to -100.0 and the second one, which will have
value 0.0 after the 'all-clocks-time-pass', will still be regarded
as a ticking (e.g. active) clock. Attending that clock will not make
the time pass, but will mantain the model semantics.

The second bug was a minor issue: long story short we now deal with
const Timeout references in the ModuleNetwork::simulation_step()
routines, as we should always have been.

TODO: clean code up, debug filth is everywhere.