Logical problem found for split importance function building.
The existential used to compute "possible rare states" per module
(in the split automatic strategy) is marking all states in the
database model as possible rare states. That's because the rare event
requires a conjunction of conditions of several modules.

A new method was deviced, where each module states' are marked as rare
iff they *definitely* contribute to a rare state.
This requires the property to be described in CNF however.

ImportanceFunctionConcrete::label_states_with_SAT() needs to be changed
completely, and the properties for all the modules need to be rewritten
in CNF.