New class for DNF property processing -- UNFINISHED.
The parsing was complex enough to merit its own class.
The new "DNFclauses" class is instantiated by ImportanceFunctionConcreteSplit
and passed on to ImportanceFunctionConcrete for the per-module importance
assessment as discussed with Pedro.

The processing of the Property by DNFclauses is almost finished,
but the ImportanceFunction construction fails nonetheless.
Review the parsing and print out the importance computed for each
ModuleInstance.