The "simulation bug" reported for the new split ifun was FIXED.
The problem was indeed in the marking of the states, but it was a logical
error regarding the *stopping* states of transient simulations; namely
the states satisfying the *negation* of the formula had to be marked
with the STOP flag (for the record, this bug was introduced last friday,
when the new per-module state marking was implemented)

TODO_0: erase all the debug filth scattered throughout the code.
TODO_1: test-check the new split ifun with all known models.
TODO_2: write the docstring of the new functions (TODO comments in place)