Fixes in the parsing and the database model.
Parsing doesn't support floating point expressions yet, only integral ones.
Real rates for the distributions have to be resolved by the user,
and the model has to be provided with the literal floating point number.
That was the fix done in the database model.

Besides, CompileModel::compileClock() was parsing the distribution rates
as integers, which resulted in exponential(0.5) parsed as exponential(0).
This led to infinite clock values and hell broke loose.
That was fixed as well.

Database model does something now, but we're not sure exactly what.
Study current behaviour.