Studying overestimation for RATE properties with RESTART engines.
Actually it just seems to affect the database model. Debug printing
with fixed-built thresholds revealed all rare events occur in the
lowermost threshold level.
That's unsurprising, since as soon as two components fail
(ImportanceValue == 2 for ifun --auto-split "+") a rare event happens.
However this may reveal some miscalculation in the weighing of
the accumulated time.

Additionally, SMC fails to build the thresholds for the database,
and the other adaptive algorithm (AMS) selects one threshold per
ImportanceValue. The only way to variate the thing is to use the
"FIX" strategy with varying splitting values.