Rate overestimation for database solved.
The problem was the identification of the RARE states during simulation.
As far as we know, whenver the rare event depends on the state of
several modules in the system, we can't store the "global rare info"
splitted per module. In those cases the rarity of a global state
during simulations needs to be decided using the property, as with
algebraic storage.

The only ifun affected by this is the ImportanceFunctionConcreteSplit.
We now let it decide whether the simulations can use its "info_of()"
member function or whether they should resort to evaluating the property.
That depends on both the property and the model, and is chosen inside the
assess_importance() method of the class.