User-defined ad hoc function extreme values -- UNFINISHED.
The ad hoc function needed for the database model uses 32 boolean variables.
Computing the extreme values of that piece of crap requires 2^32 checks,
which is stupid since we know the function can only take the values {0,1,2}

Thus we want to allow the user to specify the extreme values of the ad hoc
ImportanceFunction he provides, whenever he feels up to it.
The specification must come in the same string where he passes this ad hoc
ImportanceFunction, separated from it by semicolons (';'):
e.g. --adhoc "q2;0;9", or for the database: --adhoc "min(...);0;2".

TODO_0: finish that implementation
TODO_1: launch all experiments again in mendieta, including a new one
        for the database model