Interface change + "New" automatic function.
The functions specification names changed in the interface.
The old "--auto-coupled" is now "--amono" (as in "automatic monolithic")
The old "--auto-split"   is now "--acomp" (as in "automatic compositional")
The others ("--flat" and "--adhoc") remain unchanged,
but "--flat-coupled" and "--flat-split" have been disabled.

Two new options were added, "--amono-exp" and "--acomp-exp", equal to
those above but which store the power of two of the computed importance.
So states with regular importance 0, 1, 2... will have the "exp" importance
1, 2, 4... whenever these variants are used.

Thresholds building is a delicate issue for the "exp" variants.
For now restrict to the "fix" strategy which knows how to tackle with
big importance leaps. The other techniques should be reviewed.

A third addition in this commit is the possibility to specify user-defined
extreme values (min, max) and neutral element whenever the "--acomp" imp
function is used. These values are specified just like for the "--adhoc" imp.
function, e.g. "./fig my_model.{sa,pp} --acomp "+;0;32;0" --stop-time 1 m"
where '+' is the operand (which could be a fully-defined algebraic expression),
the first '0' is the minValue, '32' is the maxValue and the last '0' is the
neutral element to use.

TODO_0: update the experiments to use the new interface
TODO_1: review thresholds techniques to tackle with the "exp" ifun variants