Changes in hybrid thresholds builder.
When the adaptive algorithm fails and we resort to the fixed strategy
we must take the environment into account. This includes (but may not
be limited to :p) the splitting and the importance range.

Recall that's what the ThresholdsBuilderFixed does. If we instead choose
blindly a threshold every (or every other) ImportanceValue, like we were
doing so far, then an oversampling will very likely take place.

In this commit we take the first steps into including the importance range
for the fixed computation of thresholds upon a failure (or timeout) of the
adaptive algorithm. Keep investigating the matter and run tests in mendieta
to verify the efficiency obtained.