Spotted the fail causing the bug in automatic ifun construction.
This is the second BUG reported in the previous commit.
The issue is the 'varsValues_' member of the MathExpression class,
where we store the values linked to the MuParser expression.

Defining it as a vector causes trouble: delayed allocations or perhaps
reallocations yield unpredictable results when Eval()uating the expression.
We changed it to a plain C array (STATE_INTERNAL_TYPE* to be precise)
which we now have to manage by hand in ctors/copytors/dtors.

There are issues still. Declaring default copy/move ctors for MathExpression
worked but generated memfails during destruction. The current customized
versions don't work either: they behave in the same way that std::vector did.

Also regarding the first BUG reported in the previous commit,
it could've been a bad 'NVARS_' value used for std::vector::reserve()
in one of the ctors of MathExpression.

The third BUG reported wasn't studied.

TODO_0: fix MathExpression code for the "STATE_INTERNAL_TYPE* varsValues_"
TODO_1: check whether the std::bad_alloc exception BUG was fixed
TODO_2: investigate the negative clocks sampling BUG