New efficient version ready for profiling.
It was tested with --amono, --acomp, --adhoc and --flat
for default splitting (2) and thresholds (hyb) on the
tandem network model, with both debug and release builds.

The bad_alloc exception bug couldn't be reproduced,
we're leaving things as they are now.
A bug was marked in ImportanceFunction::find_extreme_values():
the minRareValue_ chosen may be fictitious. Does that affect
the computation of the estimate?

TODO_0: profile again and study new bottleneck
TODO_1: determine whether the negative clocks' sampling is
        an efficiency issue; if not, accept current solution
TODO_2: investigate the implications of a bad minRareValue_