Exponential thresholds selection -- UNFINISHED.
We want to know if the exponentiation post-processing was performed on
the ImportanceFunction, i.e. --amono-exp or --acomp-exp in the CLI,
in order to choose the thresholds "exponentially" in the non-adaptive
builders (adaptive builders handle this automatically I think)

ImportanceFunctionAlgebraic doesn't apply because there's no post-processing
offered for it. This rules out both "--adhoc" and "--flat" ifun strategies.
Hence the only concern is for the automatic ifun construction.

The approach would be to choose by importance *index* rather than *value*,
which we can do since ImportanceFunctionConcrete holds importance inside
(indexed) vectors. This abstracts away from the complexity of the values,
and should always yield true, existent importance values as threshold levels.
Piece of cake for the monolithic case, not so for the compositional case,
where the composition algebraic expression must be evaluated.

TODO_0: think of a workaround for the ImportanceFunctionConcreteSplit case
TODO_1: finish changes for ImportanceFunctionConcrete "post processing"
TODO_2: implement described changes in ThresholdsBuilder{Fixed,Hybrid}