Idea for compositional ifun: build values vector.
For the thresholds selection, instead of operating with the existent
importance vectors, we can build a container with all possible importance
values. This is the vector ImportanceFunctionConcreteCoupled already has,
and for ImportanceFunctionConcreteSplit we need to work a lot.

Is this feasible? Won't it explode in memory like the monolithic ifun?
Otherwise, how do we build the thresholds operating only with indices?