Major changes in thresholds representation.
So far we were storing the importance-to-thresholds map as we did for
bluemoon. This becomes infeasible for exponentiation post-processing,
or whenever the importance range is too big.
(I guess this is why memory usage blew up for the database experiment)

The new methodology will be to keep the inverse thresholds-to-importance
map, which should never grow bigger than ~300 locations. Then when the
threshold level of a state is asked, we us its importance to perform a
binary search on that array.

The new ThresholdsBuilder interface was defined in the base class.
Implement those changes in all the classes of the family.