Extended stop bound policies to consider relative change.
Theoretical bounds for the number of samples required by the
binomial-proportion CIs are just too hard to tune.
Either we fix them too tight and oversample (times explode),
or too lax and finish earlier but with a completely wrong estimate.

Theory (see e.g. Wikipedia) fails for extreme values of the proportion p,
which fucks us up entirely. I've been trying for the last hours to tune the
minimum number of samples and rares produced, verifying with the tandem queue
and the queue with breakdowns. All "good heuristics" proved inconsistent
when the splitting (and thus the number of thresholds) varies.

The solution found was orthogonal: besides the theoretical "static" bound,
ask for a "practical dynamic" bound based on the relative change of the
estimate, i.e. just like PRISMs epsilon between iterations.
So we *also* ask now for the difference between the new and the previous
estimates to be smaller than certain percentage of the estimate.

TODO: test in mendieta for tandem transient and qwbreaks.