Porting updates from ModelSuite to SimulationEngine  --  UNFINISHED.
New policies discussed for the simulation effort increase and for the
RATE property-type simulation.

Identify and *discard*, for steady-state properties, a first
not-steady-state behaviour:
 · simulate and increase the "simLen" effort as before;
 · in this first stage, observe outcomes but *don't update* the CI;
 · once the batch-means mechanism is "stable enough", fix simLen;
 · only from then on start updating the ConfidenceInterval.

Both the ConfidenceInterval value update and the increment of the simulation
effort are being centralized inside the SimulationEngine base class.
This leaves a cleaner ModelSuite class, and future changes in the update
policies (for both the CI and the simulations effort) should only affect
the SimulationEngine base class source file.