Change in JANI -> IOSA translation policy.
After discussions with Leo we decided the best approach is to do a
direct JANI -> ModelAST translation. Checks are performed in the
generated AST and printing to a file can be done with the already
existent ModelPrinter class.

Currently modifying ModelPrinter to make it generate a parseable
IOSA, instead of dumping debug info as it does now.