Bugfix in ImportanceFunction::Formula::operator().
There was a cast from ImportanceValue (aka size_t) to STATE_INTERNAL_TYPE
(aka short) in the evaluation of the composition function.
That created a type-size-bottleneck that returned an incorrect global
importance, namely with the MSB set.

This fixes an overflow observed in the 3tandem_queue using --acomp
with '+' as composition operand and "exp 2.83" as post processing
(or any exponent greater than 2.83).

Also polished some of Leo's code to match the coding style of the
rest of the project.