Bugfix in SimulationEngineRestart::rate_simulation().
Time truncation did not consider the existence of re-trials.

"SIM_TIME_CHUNK" was extracted from "simsLifetime" several times,
in a way that accumulated with the number of re-trials
(e.g. the aberration increased with the splitting).
Since that was not reflected in the traial.lifeTime of the next
traial, the remaining value of "simsLifetime" was incorrect.

This fix corrects the misestimation spotted by Chun in a system where
a boolean variable spends half of its time as "false" and the other
half as "true".