Extending CLI to allow specs for RNG and seed.
The task is to change everything that was decided at compile time,
into "dynamic info" that can be chosen on invocation.

Most of the code structure is finished but things got really hairy
inside Clock.cpp, due to the generalization of the "RNG type".
That type must be able to mimic C++ STL's Mersenne-Twister and also
the PCG-family RNGs, and must be usable by C++ STL's distributions.

A naïve attempt to inherint from all classes failed miserably.
Actually there's a struct with raw pointers... puke!
Ask Agus for advice?