Two bugfixes + Code refactor.
Bug #1: Leo forgot to update the value of the clocks in the module where
        the output-committed action is taken.

Bug #2: Leo broke the committed-checking-loop as soon as a single module
        didn't have committed actions.

Both bugs appeared in ModuleNetwork::process_committed_once().
A refactoring of the whole treatment for committed actions,
resembling the treatment of timed actions and affecting the
ModuleNetwork and ModuleInstance classes, fixed both issues.