Bugfix spotted by Chun.
When performing a simulation step, the process_committed() must be run
every time *right before* checking the loop's guard.
This was not so, since process_committed() was invoked after checking
the loop's guard in ModuleNetwork::simulation_step().

Imitating the algorithm previously fixed in ModuleNetwork::peak_simulation()
solved the issue. Well done Chun!

This fix corrects the misestimation of the AND-gate-system spotted by Chun,
where observing the variables of the BEs yielded the correct result,
but observing the variable of the AND gate yielded a wrong number.