Bugfix by Raúl + Code structure enhancement.
Raúl got rid of the bug formerly spotted by Chun,
changing the layout of the passive jump in ModuleInstance.
Now we "kill time" uniformly in all the clocks of the current module,
and _then_ reset the clocks of the matching enabled transition, if any.

Moreover, these updates triggered cascading modularisations and TODOs
left behind by Leo.
These issues have been taken care of, and the general structure of the
code has improved, thanks to Raúl :D

BTW, the code in the files touched is filthy.   Cleanse ASAP.