Update in tandem queue test case.
We already tested the behaviour of standard Monte Carlo simulations,
and RESTART simulations with ad hoc and automatic monolithic ifuns,
for both transient and steady state properties.

Right now stuck with RESTART + compositional ifun for steady-state query:
the last SECTION of the TEST_CASE is failing with SIGSEV.

-----------------------------------
The bug causing this may be broader

Since our models are not parametric, we rely on the values hard-coded
into the model files. For instance the capacity 'c' of the tandem queue
has to be equal to 10 in the model file, otherwise the tests fail.

Is there an easy workaround?