Model files for the test suite.
The easiest workaround to not having parametric models
is copying here all the model we want to test.
Thus we now a have a dedicated tests/models subdir,
where all model files used for testing must be located.

This also allows defining in the files all properties we will test,
without the risk of a test-suite-breakdown if someone fiddles with
the model files in the main "models" directory of the repo.